requirements.txt
pyserial
pyyaml
numpy
scikit-learn
matplotlib
PyQt6
